Add index.html instead of ignoring it
In a previous version of our code, we used a homepage.html which
we built using Make into index.html. We had git ignore the index.html
because it was a built file. Now, however, we are using Three.js,
and we write the index.html directly. We want git to track it.