Improve accessibility for the blind and hard-of-seeing
Every select tag should have an associated label. When
visually-impaired people use their computer, they can enable a
screen reader. The reader would go through each element on the HTML
page and read it out loud. Without a label, it would not know what
to call a select tag, so the user might not understand what is on
the screen.

We only have one user in mind, the grader, who is probably not
blind. Also, the grader is probably not grading based on
accessibility, and the other groups probably do not care. However,
it's important to make software that supports everyone, including
the disabled.

There's more information here:
https://dequeuniversity.com/rules/axe/4.1/select-name