Add custom bodyparts on click
The code reads the custom, more complex bodyparts from the server.
I had to replace the multiple selects with a single select for
the body part because the code assumes that there is only one, and
it would take too long to change the code to support several selects.
If we had more time, then I would keep the multiple selects.

Note that the custom body parts are reloaded from the server every
time one is made, so it is slow. We need to implement caching,
especially because the hover creates and destroys bodyparts quickly.

Also, the horse torso does not load in this commit. We need to fix
that.