Cache body parts
If the user wants to add an eye twice, then we only need to load
the eye once from the server. The next time, we can reuse the object
we downloaded.

This also makes preview on hover possible because the preview creates
and destroys the previewed body part as the user moves the mouse.